split profile component, moved profile
